requirements.txt
imageio==2.26.0
matplotlib==3.7.0
numexpr==2.8.4
numpy==1.24.2
onnx==1.13.1
onnx_simplifier==0.4.17
onnxruntime==1.14.1
onnxsim==0.4.17
opencv_python==4.7.0.72
pandas==1.5.3
Pillow==9.4.0
pyclipper==1.3.0.post4
python-dotenv==1.0.0
PyYAML==6.0
requests==2.28.2
scipy==1.10.1
setuptools==67.6.0
shapely==2.0.1
skimage==0.0
torch==1.13.1
torchsummary==1.5.1
torchvision==0.14.1
tqdm==4.65.0
trdg==1.8.0
wandb==0.13.10
